refactor(policy): centralize shared helpers
