+++ extensions/policy/src/policy-state-workspace.ts
@@ -3,12 +3,9 @@ import {
   isRecord,
   normalizeOptionalString as readString,
 } from "openclaw/plugin-sdk/string-coerce-runtime";
-import {
-  AGENT_WORKSPACE_POLICY_TOOLS,
-  readStringArray,
-  toolListCoversTool,
-} from "./policy-state-tool-posture.js";
+import { AGENT_WORKSPACE_POLICY_TOOLS, readStringArray } from "./policy-state-tool-posture.js";
 import type { PolicyAgentWorkspaceEvidence } from "./policy-state-types.js";
+import { toolListCoversTool } from "./tool-policy-conformance.js";
 
 export function scanPolicyAgentWorkspace(
   cfg: Record<string, unknown>,
